less(1): Fix type mismatches in different objects.
Not sure which one is better int or long. For now downgrade to int,
cause it is less changes in contrib code.

+++ contrib/less/opttbl.c
@@ -39,8 +39,8 @@ public int ctldisp;		/* Send control chars to screen untranslated */
 public int force_open;		/* Open the file even if not regular file */
 public int swindow;		/* Size of scrolling window */
 public int jump_sline;		/* Screen line of "jump target" */
-public long jump_sline_fraction = -1;
-public long shift_count_fraction = -1;
+public int jump_sline_fraction = -1;
+public int shift_count_fraction = -1;
 public int chopline;		/* Truncate displayed lines at screen width */
 public int no_init;		/* Disable sending ti/te termcap strings */
 public int no_keypad;		/* Disable sending ks/ke termcap strings */
